requirements.txt
# Production Runtime Dependencies
# Optimized for CPU-only inference, minimal Docker image size

# Core Flask (Web Framework)
Flask==3.0.2
Werkzeug==3.0.1
gunicorn==23.0.0

# Flask Extensions
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0

# Deep Learning - CPU Only
# Install with: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
# Using standard pip will auto-select CPU wheels on Linux
torch==2.2.0
torchvision==0.17.0

# Image/Video Processing (headless = no GUI dependencies)
opencv-python-headless==4.9.0.80
Pillow==10.2.0

# Database
SQLAlchemy==2.0.25

# Note: This configuration targets <1.5GB Docker image
# PyTorch CPU wheels: ~750MB, OpenCV: ~90MB, other deps: ~150MB
# Total: ~1GB + 216MB model + base image = ~1.4GB final size